디렉토리 이름 변경 기존 vibestay > webcore 로 변경함 데이터 베이스 이름 변경 vibestay > platform
변경 전 버전은 03.26 버전입니다.

- bongJung 2025-04-24 -
