--- src/main/java/com/vibeStay/platform/PlatformApplication.java
@@ -1,13 +1,0 @@
-package com.vibeStay.platform;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PlatformApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(PlatformApplication.class, args);
-	}
-
-}
