--- src/main/java/com/vibeStay/platform/domain/AuthenticationRequest.java
@@ -1,10 +1,0 @@
-package com.vibeStay.platform.domain;
-
-import lombok.Data;
-
-@Data
-public class AuthenticationRequest {
-    private String userId;
-    private String password;
-
-}
