+++ src/main/java/com/webcore/platform/PlatformApplication.java
@@ -1,0 +1,13 @@
+package com.webcore.platform;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PlatformApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PlatformApplication.class, args);
+	}
+
+}
