+++ src/main/java/com/webcore/platform/domain/AuthenticationRequest.java
@@ -1,0 +1,10 @@
+package com.webcore.platform.domain;
+
+import lombok.Data;
+
+@Data
+public class AuthenticationRequest {
+    private String userId;
+    private String password;
+
+}
