+++ src/main/resources/application.properties
@@ -5,11 +5,11 @@ server.port=8080
 spring.mvc.view.prefix=/WEB-INF/views/
 spring.mvc.view.suffix=.jsp
 
-com.vibeStay.platform.secret-key=lviNqkZQtVXxRDf3VVNCSALLLqMgrmKQ9fCFhHcA24Ep4jgqyxkT21gxvCJTmhZZ
+com.webcore.platform.secret-key=lviNqkZQtVXxRDf3VVNCSALLLqMgrmKQ9fCFhHcA24Ep4jgqyxkT21gxvCJTmhZZ
 
 # MySQL
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://localhost:3308/vibestay
+spring.datasource.url=jdbc:mysql://localhost:3308/platfrom
 spring.datasource.username=root
 spring.datasource.password=1234
 
@@ -18,5 +18,5 @@ mybatis.mapper-locations=classpath:mybatis/mapper/**/**.xml
 # MyBatis
 mybatis.configuration.map-underscore-to-camel-case=true
 # Mybatis ResultType
-mybatis.type.aliases-package=com.vibeStay.platform.domain
+mybatis.type.aliases-package=com.webcore.platform.domain
