--- src/test/java/com/vibeStay/platform/PlatformApplicationTests.java
@@ -1,13 +1,0 @@
-package com.vibeStay.platform;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PlatformApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
