+++ src/test/java/com/webcore/platform/PlatformApplicationTests.java
@@ -1,0 +1,13 @@
+package com.webcore.platform;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PlatformApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
